requirement.txt
opencv-python
scikit-image
pillow
matplotlib
hdf5storage
ninja
lmdb
requests
timm
einops
torchinfo
lpips